% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ipc_articulos_2020_2024}
\alias{data_ipc_articulos_2020_2024}
\title{CPI by item data, from October 2020 to December 2024}
\format{
\subsection{\code{data_ipc_articulos_2020_2024}}{

A data frame with 612 rows and 26 columns:

\describe{
\item{id}{Order in the original excel}
\item{nombre}{Item name, it can be a group, class or other level}
\item{agregacion}{This is the level of the given item}
\item{group}{Group name}
\item{subgrupo}{Subgroup name}
\item{clase}{Class name}
\item{subclase}{Subclass name}
\item{articulo}{Item name}
\item{articulo}{Item name}
\item{date}{Date}
\item{year}{Year}
\item{mes}{Month}
\item{ponderacion_ipc}{Weight of the element as proportion of the general basket}
\item{indice}{item index}
}
}
}
\usage{
data_ipc_articulos_2020_2024
}
\description{
CPI by item data, from October 2020 to December 2024
}
\keyword{datasets}
