% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_label}
\alias{date_label}
\title{Create date label}
\usage{
date_label(date = Sys.Date())
}
\arguments{
\item{date}{a vector of class Date}
}
\value{
a character string
}
\description{
Create label in "Month Year" format out of a date object
}
