% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasas_sf.R
\name{.download_bcrd_file}
\alias{.download_bcrd_file}
\title{Descargador Centralizado de Archivos BCRD (Punto 1 y 4)}
\usage{
.download_bcrd_file(url, dest_path)
}
\description{
Descargador Centralizado de Archivos BCRD (Punto 1 y 4)
}
