% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasas_sf.R
\name{.get_historical_rates}
\alias{.get_historical_rates}
\title{Motor de extracción unificado para series mensuales históricas (Punto 1 y 6)}
\usage{
.get_historical_rates(type = c("pasiva", "activa"))
}
\description{
Motor de extracción unificado para series mensuales históricas (Punto 1 y 6)
}
