% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expectativas_details}
\alias{expectativas_details}
\title{Utility complement to handle the macroeconomic expectations}
\format{
\subsection{\code{expectativas_details}}{

A data frame with 39 rows and 6 columns:
}
}
\usage{
expectativas_details
}
\description{
Utility complement to handle the macroeconomic expectations
}
\keyword{datasets}
