% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectativas.R
\name{get_expectativas}
\alias{get_expectativas}
\title{Macroeconomic Expectations}
\usage{
get_expectativas(modalidad = "eem")
}
\arguments{
\item{modalidad}{string with the desired disaggregation. options:
"eem", "eoe", "ecc",}
}
\value{
a tibble
}
\description{
Get the data series for the monthly macroeconomic expectations of the
Dominican Republic.
}
\examples{
get_expectativas("eem")
get_expectativas("eoe")
get_expectativas("ecc")

}
