% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbkf.R
\name{get_fbkf}
\alias{get_fbkf}
\title{Download GROSS FIXED CAPITAL FORMATION BY SECTOR AND TYPE OF GOOD}
\usage{
get_fbkf()
}
\value{
a data frame with monthly series
}
\description{
Get the series of the gross fixed capital formation by sector and
type of good of the Dominican Republic.
}
\examples{
get_fbkf()
}
