% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal.R
\name{get_fiscal}
\alias{get_fiscal}
\title{Download Statement of Operations of the non-financial public sector}
\usage{
get_fiscal(frecuencia = "Mensual")
}
\arguments{
\item{frecuencia}{A character string that specifies the frequency of the
exchange rates to be downloaded. Valid options are "Mensual",  or "Anual".}
}
\value{
a tibble
}
\description{
Download Statement of Operations of the non-financial public sector
}
\examples{
get_fiscal()
}
