% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ied.R
\name{get_ied}
\alias{get_ied}
\title{Foreign direct investment to the Dominican Republic}
\usage{
get_ied(modalidad = "pais_origen", frecuencia = "trimestral")
}
\arguments{
\item{modalidad}{A character string. Options are "pais_origen" or
"sector_destino".}

\item{frecuencia}{A character string that specifies the frequency of the
exchange rates to be downloaded. Valid options are "trimestral",  or "anual".}
}
\value{
A data frame
}
\description{
This function returns the  Foreign direct investment to the
This function returns the  Foreign direct investment to the
Dominican Republic based on the specified frequency.
}
\examples{
get_tc_eif("mensual")
get_tc_eif("anual")
}
