% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_bcrd.R
\name{get_indicadores_monetarios_bcrd}
\alias{get_indicadores_monetarios_bcrd}
\title{Monetary Indicators of the Central Bank of the Dominican Republic}
\usage{
get_indicadores_monetarios_bcrd()
}
\value{
a data frame
}
\description{
Monetary Indicators of the Central Bank of the Dominican Republic
}
