% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_articulos.R
\name{get_ipc_articulos}
\alias{get_ipc_articulos}
\title{Download the CPI series at item level}
\usage{
get_ipc_articulos()
}
\value{
a tibble
}
\description{
Download the CPI inflacion series for the Dominican Republic by item level with details
about the group, subgroup, class and subclass
}
\examples{
get_ipc_articulos("general")
}
