% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_articulos.R
\name{get_ipc_long}
\alias{get_ipc_long}
\title{Download the CPI series in long format}
\usage{
get_ipc_long(
  desagregacion = c("general", "grupo", "subgrupo", "clase", "subclase", "articulo")
)
}
\arguments{
\item{desagregacion}{string with the desired disaggregation. options:
"general", "grupo", "subgrupo", "clase", "subclase", "articulo"}
}
\value{
a tibble
}
\description{
Download the CPI inflacion series for the Dominican Republic in any
disaggregation
}
\examples{
get_ipc_long("general")
get_ipc_long("grupo")
}
