% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib_gasto.R
\name{get_pib_gasto}
\alias{get_pib_gasto}
\title{GDP by the expenditure approach, both real and nominal}
\usage{
get_pib_gasto(modalidad = "nominal", acumulado = FALSE, homogenea_91 = FALSE)
}
\arguments{
\item{modalidad}{string indicating if "real" or "nominal"}

\item{acumulado}{logical, by default FALSE}

\item{homogenea_91}{logical, by default FALSE}
}
\value{
a data frame
}
\description{
GDP by the expenditure approach, both real and nominal
}
