% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib_sectores.R
\name{get_pib_sectores}
\alias{get_pib_sectores}
\title{GDP by the sector of origin, both real and nominal}
\usage{
get_pib_sectores(modalidad = "real", acumulado = FALSE, homogenea_91 = FALSE)
}
\arguments{
\item{modalidad}{string indicating if "real" or "nominal"}

\item{acumulado}{logical, by default FALSE}

\item{homogenea_91}{logical, by default FALSE}
}
\value{
a data frame
}
\description{
GDP by the sector of origin, both real and nominal
}
