% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestamos_sectores.R
\name{get_prestamos_osd}
\alias{get_prestamos_osd}
\title{Download the loans by sector series}
\usage{
get_prestamos_osd(osd = "consolidado")
}
\arguments{
\item{osd}{string with the desired societies options:
"consolidado", "bancos_multiples", "resto_osd".}
}
\value{
a tibble
}
\description{
Download loans by sectors
}
\examples{
get_prestamos_osd("consolidado")
get_prestamos_osd("bancos_multiples")
get_prestamos_osd("resto_osd")
}
