% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remesas.R
\name{get_remesas}
\alias{get_remesas}
\title{Remittances to the Dominican Republic}
\usage{
get_remesas(modalidad = "mensual")
}
\arguments{
\item{modalidad}{string indicating which perspective of remittances is asked.
Options are: "mensual", "por_pais_emisor", "por_provincia_receptora",
"cantidad_de_transacciones", "promedio_transacciones", "segun_moneda",
"entidad_pagadora", "genero_receptor"}
}
\value{
a data frame
}
\description{
Remittances to the Dominican Republic
}
\examples{
get_remesas("mensual")
get_remesas("por_pais_emisor")
get_remesas("por_provincia_receptora")
get_remesas("cantidad_de_transacciones")
get_remesas("promedio_transacciones")
get_remesas("segun_moneda")
get_remesas("entidad_pagadora")
get_remesas("genero_receptor")
}
