% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasas_sf.R
\name{get_tasas_activas}
\alias{get_tasas_activas}
\title{Download interest rates for loans}
\usage{
get_tasas_activas()
}
\value{
a data frame with monthly series
}
\description{
Get the series of the interest rates for loans in the dominican republic
separated by time and type of loan
}
