% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo_cambio.R
\name{get_tc}
\alias{get_tc}
\title{Average exchange rate of the operations in financial institutions}
\usage{
get_tc(frecuencia = "diaria", entidad = "spot", average_or_fp = "average")
}
\arguments{
\item{frecuencia}{A character string that specifies the frequency of the
exchange rates to be downloaded. Valid options are "diaria", "mensual",
"trimestral",  or "anual".}

\item{entidad}{Valid options are "eif" for financial institutions,
"ac" for foreign currency exchange agencies or "spot" for reference
exchange rate}

\item{average_or_fp}{A character string that specifies if the average or the
value for the last day of the period is desired. valid options are "average"
and "fp" for end of period.}
}
\value{
A data frame with columns:
compra: for the buying rates
venta: selling rates
}
\description{
This function returns the average exchange rates of the operations conducted
by the banks and exchange operators in the  Dominican Republic
based on the specified frequency.
}
\examples{
get_tc(frecuencia = "diaria", entidad = "spot", average_or_fp = "average")
get_tc(frecuencia = "mensual", entidad = "eif", average_or_fp = "average")
get_tc(frecuencia = "trimestral", entidad = "ac", average_or_fp = "fp")
}
