% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo_cambio.R
\name{get_tc_eif}
\alias{get_tc_eif}
\title{Average exchange rate of the operations in financial institutions}
\usage{
get_tc_eif(frecuencia = "diaria")
}
\arguments{
\item{frecuencia}{A character string that specifies the frequency of the
exchange rates to be downloaded. Valid options are "diaria", "mensual",
"trimestral",  or "anual".}
}
\value{
A data frame with columns:
compra: for the buying rates
venta: selling rates
}
\description{
This function returns the average exchange rates of the operations conducted
by the banks and exchange operators in the  Dominican Republic
based on the specified frequency.
}
\examples{
get_tc_eif("mensual")
get_tc_eif("anual")
}
