% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo_cambio.R
\name{get_tc_spot}
\alias{get_tc_spot}
\title{Exchange rates in the spot market of the Dominican Republic}
\usage{
get_tc_spot(frecuencia = "mensual", average_or_fp = "average")
}
\arguments{
\item{frecuencia}{A character string that specifies the frequency of the
exchange rates to be downloaded. Valid options are "diaria", "mensual",
"trimestral",  or "anual".}

\item{average_or_fp}{A character string that specifies if the average or the
value for the last day of the period is desired. valid options are "average"
and "fp"}
}
\value{
A data frame with columns:
compra: for the buying rates
venta: selling rates
}
\description{
This function returns the average exchange rates of the operations conducted
by the banks and exchange operators in the  Dominican Republic
based on the specified frequency.
}
\examples{
get_tc_spot("mensual", "average")
get_tc_spot("mensual", "fp")
get_tc_spot("trimestral", "average")
get_tc_spot("trimestral", "fp")
}
