% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.R
\name{get_tpm}
\alias{get_tpm}
\title{Series of the Monetary Policy Rate of the Dominican Republic}
\usage{
get_tpm()
}
\value{
a tibble
}
\description{
Monetary Policy Rate as well as some short term rates for deposits or loans
}
