% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicadores_bcrd_details}
\alias{indicadores_bcrd_details}
\title{Utility complement to handle the monetary indicators data}
\format{
\subsection{\code{indicadores_bcrd_details}}{

A data frame with 39 rows and 6 columns:
\describe{
\item{original_names}{Names in the excel file}
\item{labels}{Names with meaning and without suffix}
\item{short_names}{unique names in snake case}
\item{categoria}{Category of the indicator}
\item{nivel}{level}
\item{direct_parent}{Direct parent}
}
}
}
\usage{
indicadores_bcrd_details
}
\description{
Utility complement to handle the monetary indicators data
}
\keyword{datasets}
