% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc_articulos_details}
\alias{ipc_articulos_details}
\title{Utility complement to handle the CPI by item}
\format{
\subsection{\code{ipc_articulos_details}}{

A data frame with 612 rows and 26 columns:

\describe{
\item{posicion}{Order in the original excel}
\item{nombre}{Item name, it can be a group, class or other level}
\item{agregacion}{This is the level of the given item}
\item{group}{Group name}
\item{subgrupo}{Subgroup name}
\item{clase}{Class name}
\item{subclase}{Subclass name}
\item{articulo}{Item name}
\item{is_grupo}{Boolean indicating if is a group}
\item{is_subgrupo}{Boolean indicating if is a subgroup}
\item{is_clase}{Boolean indicating if is a class}
\item{is_subclase}{Boolean indicating if is a subclass}
\item{is_articulo}{Boolean: is an item?}
\item{is_subyacente}{Boolean: is part of the core inflation?}
\item{is_no_subyacente}{Boolean: is part of the non-core inflation items?}
\item{is_transables}{Boolean: is a tradable good?}
\item{is_no_transables}{Boolean: is a non-tradable good?}
\item{is_bienes}{Boolean: is a good?}
\item{is_servicios}{Boolean: is a service?}
\item{is_alimentos}{Boolean: is a food item?}
\item{is_transporte}{Boolean: is a transportation item?}
\item{is_vivienda}{Boolean: is housing?}
\item{is_resto}{Boolean: is others?}
\item{is_nuevo}{Boolean: was introduced in the most recent basket?}
\item{codigo_articulo}{id of the item}
\item{ponderacion_ipc}{weight of the element as proportion of the general basket}
}
}
}
\usage{
ipc_articulos_details
}
\description{
Utility complement to handle the CPI by item
}
\keyword{datasets}
