% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detalles.R
\name{read_details}
\alias{read_details}
\title{Function to read files metadata}
\usage{
read_details(topic = NULL)
}
\arguments{
\item{topic}{specific topic to query}
}
\value{
a list
}
\description{
Function to read files metadata
}
