% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{change the scale of a vector}
\usage{
rescale(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{new_min}{numeric scalar with the minimum value desired}

\item{new_max}{numeric scalar with the maximum value desired}
}
\value{
a numeric vector
}
\description{
With this function you can change the scale of a numeric vector providing the
new minimum and maximum values
}
\examples{
rescale(1:10)
}
