% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasas_sf.R
\name{tasas_to_long}
\alias{tasas_to_long}
\title{Pivot wide formats of interest rates data frames to long format}
\usage{
tasas_to_long(
  tasas_wide,
  filtro_tipo_tasa = NULL,
  filtro_moneda = NULL,
  filtro_condicion = NULL,
  filtro_grupo = NULL,
  filtro_detalle = NULL
)
}
\arguments{
\item{tasas_wide}{\verb{<data.frame>} Wide structured table array parsed from source files.}

\item{filtro_tipo_tasa}{\verb{<character>} Filter by type: \code{"Activa"} or \code{"Pasiva"}.}

\item{filtro_moneda}{\verb{<character>} Filter by currency: \code{"DOP"} or \code{"USD"}.}

\item{filtro_condicion}{\verb{<character>} Filter by rate condition: \code{"General"} or \code{"Preferencial"}.}

\item{filtro_grupo}{\verb{<character>} Filter by grouping: \code{"Plazo"}, \code{"Promedio"}, or \code{"Sector"}.}

\item{filtro_detalle}{\verb{<character>} Filter by specific descriptive category.}
}
\value{
A tidy mapped structure long configuration \code{tibble}.
}
\description{
Tidies interest rate datasets from wide matrix forms down into structured tidy key-value combinations.
}
