% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zona_euro.R
\name{ze_inflation}
\alias{ze_inflation}
\title{Download euro area inflation data}
\usage{
ze_inflation()
}
\value{
A tibble with two columns:
\describe{
\item{date}{Observation date.}
\item{inflation}{Monthly HICP inflation rate.}
}
}
\description{
Retrieves the monthly Harmonised Index of Consumer Prices (HICP) inflation
series for the euro area from the European Central Bank (ECB).
}
\details{
The data correspond to the ECB series
\code{"HICP.M.U2.N.000000.4D0.ANR"}, which reports the annual rate of change of
the Harmonised Index of Consumer Prices (HICP) for the euro area.
}
\seealso{
Obtain additional information about the ECB Data Portal at
\url{https://data.ecb.europa.eu/}.
}
