% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exports.R
\name{get_exportaciones}
\alias{get_exportaciones}
\title{Total exports by sectors}
\usage{
get_exportaciones(frecuencia = "mensual")
}
\arguments{
\item{frecuencia}{A character string that specifies the frequency of the
data to be downloaded. Valid options are "mensual",
"trimestral",  or "anual".}
}
\value{
A data frame
}
\description{
This function returns total exports by sectors in the  Dominican Republic
based on the specified frequency.
}
\examples{
get_exportaciones("mensual")
get_exportaciones("trimestral")
get_exportaciones("anual")
}
