% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exports.R
\name{get_exportaciones_zf}
\alias{get_exportaciones_zf}
\title{Free Trade Exports}
\usage{
get_exportaciones_zf()
}
\value{
A data frame
}
\description{
This function returns Free Trade exports from the Dominican Republic
by goods
}
\examples{
get_exportaciones_zf()
}
