% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exports.R
\name{get_importaciones}
\alias{get_importaciones}
\title{Total imports by sectors}
\usage{
get_importaciones(frecuencia = "mensual")
}
\arguments{
\item{frecuencia}{A character string that specifies the frequency of the
data to be downloaded. Valid options are "mensual",
"trimestral",  or "anual".}
}
\value{
A data frame
}
\description{
This function returns total imports by sectors in the  Dominican Republic
based on the specified frequency.
}
\examples{
get_importaciones("mensual")
get_importaciones("trimestral")
get_importaciones("anual")
}
