% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exports.R
\name{get_importaciones_petroleo}
\alias{get_importaciones_petroleo}
\title{Oil Imports}
\usage{
get_importaciones_petroleo()
}
\value{
A data frame
}
\description{
This function returns oil imports to the Dominican Republic
by type
}
\examples{
get_exportaciones_zf()
}
