% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{get_ipc_data}
\alias{get_ipc_data}
\title{Download the CPI series}
\usage{
get_ipc_data(desagregacion = "general")
}
\arguments{
\item{desagregacion}{string with the desired disaggregation. options:
"general", "grupos", "regiones", "subyacente", "tnt" (transable y no transable),
"articulos"}
}
\value{
a tibble
}
\description{
Download the CPI inflacion series for the Dominican Republic in any
disaggregation
}
\examples{
get_ipc_data()
}
