% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{get_ipc_general}
\alias{get_ipc_general}
\title{To get the general CPI data}
\usage{
get_ipc_general()
}
\description{
You can get the CPI index and the monthly, year over year and
throughout the year variations, as well as the 12 month average
}
