% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{get_ipc_subyacente}
\alias{get_ipc_subyacente}
\title{To get the CPI data by group of geographic region}
\usage{
get_ipc_subyacente()
}
\description{
You can get the CPI index and the monthly, year over year and
throughout the year variations, as well as the 12 month average
}
