% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{get_ipc_tnt}
\alias{get_ipc_tnt}
\title{To get the CPI data by transferable or not}
\usage{
get_ipc_tnt()
}
\description{
You can get the CPI index and the monthly, year over year and
throughout the year variations, as well as the 12 month average
}
