% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasas_sf.R
\name{get_tasas_pasivas}
\alias{get_tasas_pasivas}
\title{Download interest rates for savings}
\usage{
get_tasas_pasivas()
}
\value{
a data frame with monthly series
}
\description{
Get the series of the interest rates for savings in the Dominican Republic
separated by time and type of saving
}
